+++ react-native-firebase-starter/android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.invertase.rnfirebasestarter",
+    package = "com.none.firebasetemplate",
 )
 
 android_resource(
     name = "res",
-    package = "com.invertase.rnfirebasestarter",
+    package = "com.none.firebasetemplate",
     res = "src/main/res",
 )
