+++ react-native-firebase-starter/android/app/src/main/java/com/none/firebasetemplate/MainActivity.java
@@ -1,4 +1,4 @@
-package com.invertase.rnfirebasestarter;
+package com.none.firebasetemplate;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "RNFirebaseStarter";
+        return "FireBaseTemplate";
     }
 }
