+++ react-native-firebase-starter/android/app/src/main/java/com/none/firebasetemplate/MainApplication.java
@@ -1,4 +1,4 @@
-package com.invertase.rnfirebasestarter;
+package com.none.firebasetemplate;
 
 import android.annotation.SuppressLint;
 import android.app.Application;
