+++ react-native-firebase-starter/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "RNFirebaseStarter",
-  "displayName": "RNFirebaseStarter"
+  "name": "FireBaseTemplate",
+  "displayName": "FireBaseTemplate"
 }
