+++ react-native-firebase-starter/ios/FireBaseTemplate-tvOS/Info.plist
