+++ react-native-firebase-starter/ios/FireBaseTemplate-tvOSTests/Info.plist
