+++ react-native-firebase-starter/ios/FireBaseTemplate.xcodeproj/xcshareddata/xcschemes/FireBaseTemplate.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "RNFirebaseStarter.app"
-               BlueprintName = "RNFirebaseStarter"
-               ReferencedContainer = "container:RNFirebaseStarter.xcodeproj">
+               BuildableName = "FireBaseTemplate.app"
+               BlueprintName = "FireBaseTemplate"
+               ReferencedContainer = "container:FireBaseTemplate.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "RNFirebaseStarterTests.xctest"
-               BlueprintName = "RNFirebaseStarterTests"
-               ReferencedContainer = "container:RNFirebaseStarter.xcodeproj">
+               BuildableName = "FireBaseTemplateTests.xctest"
+               BlueprintName = "FireBaseTemplateTests"
+               ReferencedContainer = "container:FireBaseTemplate.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "RNFirebaseStarterTests.xctest"
-               BlueprintName = "RNFirebaseStarterTests"
-               ReferencedContainer = "container:RNFirebaseStarter.xcodeproj">
+               BuildableName = "FireBaseTemplateTests.xctest"
+               BlueprintName = "FireBaseTemplateTests"
+               ReferencedContainer = "container:FireBaseTemplate.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNFirebaseStarter.app"
-            BlueprintName = "RNFirebaseStarter"
-            ReferencedContainer = "container:RNFirebaseStarter.xcodeproj">
+            BuildableName = "FireBaseTemplate.app"
+            BlueprintName = "FireBaseTemplate"
+            ReferencedContainer = "container:FireBaseTemplate.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNFirebaseStarter.app"
-            BlueprintName = "RNFirebaseStarter"
-            ReferencedContainer = "container:RNFirebaseStarter.xcodeproj">
+            BuildableName = "FireBaseTemplate.app"
+            BlueprintName = "FireBaseTemplate"
+            ReferencedContainer = "container:FireBaseTemplate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNFirebaseStarter.app"
-            BlueprintName = "RNFirebaseStarter"
-            ReferencedContainer = "container:RNFirebaseStarter.xcodeproj">
+            BuildableName = "FireBaseTemplate.app"
+            BlueprintName = "FireBaseTemplate"
+            ReferencedContainer = "container:FireBaseTemplate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
