+++ react-native-firebase-starter/ios/FireBaseTemplate.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:RNFirebaseStarter.xcodeproj">
+      location = "group:FireBaseTemplate.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
