+++ react-native-firebase-starter/ios/FireBaseTemplate.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
