+++ react-native-firebase-starter/ios/FireBaseTemplate.xcworkspace/xcshareddata/WorkspaceSettings.xcsettings
