+++ react-native-firebase-starter/ios/FireBaseTemplate/AppDelegate.h
