+++ react-native-firebase-starter/ios/FireBaseTemplate/Images.xcassets/AppIcon.appiconset/Contents.json
