+++ react-native-firebase-starter/ios/FireBaseTemplate/Info.plist
