+++ react-native-firebase-starter/ios/FireBaseTemplate/main.m
