+++ react-native-firebase-starter/ios/FireBaseTemplateTests/Info.plist
