+++ react-native-firebase-starter/ios/Podfile
@@ -1,11 +1,11 @@
 # Uncomment the next line to define a global platform for your project
 platform :ios, '9.0'
 
-target 'RNFirebaseStarter' do
+target 'FireBaseTemplate' do
   # Uncomment the next line if you're using Swift or would like to use dynamic frameworks
   # use_frameworks!
 
-  # Pods for RNFirebaseStarter
+  # Pods for FireBaseTemplate
 
   # Required by RNFirebase
   pod 'Firebase/Core', '~> 5.11.0'
