+++ react-native-firebase-starter/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "RNFirebaseStarter",
+  "name": "FireBaseTemplate",
   "version": "5.1.0",
   "private": true,
   "devDependencies": {
